Extracted TestLogger from SandboxSpecs so it can be used elsewhere
